+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updates invalid text in the `SearchBar` component to include "There was a problem. " prefix.
+- Adds interaction tests to `MultiSelect` component.
+- Increases the z-index of the `FeedbackBox`'s `Drawer` component so it displays above all other elements on a page.
 
 ## 3.3.2 (September 19, 2024)